+++ hydra-node/test/Hydra/Logging/MonitoringSpec.hs
@@ -15,7 +15,7 @@ import Hydra.Logging (nullTracer, traceWith)
 import Hydra.Logging.Messages (HydraLog (Node))
 import Hydra.Logging.Monitoring
 import Hydra.Network.Message (Message (ReqTx))
-import Hydra.Node (HydraNodeLog (ProcessedEffect, ProcessingEvent))
+import Hydra.Node (HydraNodeLog (BeginEvent, EndEffect))
 import Hydra.ServerOutput (ServerOutput (SnapshotConfirmed))
 import Hydra.Snapshot (Snapshot (Snapshot))
 import Network.HTTP.Req (GET (..), NoReqBody (..), bsResponse, defaultHttpConfig, http, port, req, responseBody, runReq, (/:))
@@ -28,10 +28,10 @@ spec = do
     failAfter 3 $ do
       [p] <- randomUnusedTCPPorts 1
       withMonitoring (Just $ fromIntegral p) nullTracer $ \tracer -> do
-        traceWith tracer (Node $ ProcessingEvent alice (NetworkEvent (ReqTx alice (aValidTx 42))))
-        traceWith tracer (Node $ ProcessingEvent alice (NetworkEvent (ReqTx alice (aValidTx 43))))
+        traceWith tracer (Node $ BeginEvent alice (NetworkEvent (ReqTx alice (aValidTx 42))))
+        traceWith tracer (Node $ BeginEvent alice (NetworkEvent (ReqTx alice (aValidTx 43))))
         threadDelay 0.1
-        traceWith tracer (Node $ ProcessedEffect alice (ClientEffect (SnapshotConfirmed (Snapshot 1 (utxoRefs [1]) [aValidTx 43, aValidTx 42]) mempty)))
+        traceWith tracer (Node $ EndEffect alice (ClientEffect (SnapshotConfirmed (Snapshot 1 (utxoRefs [1]) [aValidTx 43, aValidTx 42]) mempty)))
 
         metrics <-
           Text.lines . decodeUtf8 . responseBody